+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added flow to verify account address.
 * Added support for register data transaction.
+* Support for the configure delegation transaction.
 
 ## 2.0.3 - 2021-11-29